# build/openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/request/templates/{templateId}": {
      "get": {
        "operationId": "getRequestTemplate",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Got request templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestTemplateResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get request templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Get a requests template",
        "tags": [
          "requestTemplates"
        ]
      }
    },
    "/request/templates": {
      "get": {
        "operationId": "getRequestTemplates",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Template type \"service_request\", \"parts_request\"",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Got request templatess",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestTemplateListReponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get request templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Get all requests template",
        "tags": [
          "requestTemplates"
        ]
      }
    },
...
    "/profiles/requester/request": {
      "post": {
        "operationId": "createRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create a request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Create a request for the user ",
        "tags": [
          "profiles",
          "requests"
        ]
      }
    },
    "/profiles/requester/request/{id}": {
      "patch": {
        "operationId": "updateRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create a request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Update a request for the user ",
        "tags": [
          "profiles",
          "requests"
        ]
      },
      "get": {
        "operationId": "getRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Got request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Failed to get a request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get a request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Get a request for the user ",
        "tags": [
          "profiles",
          "requests"
        ]
      },
      "delete": {
        "operationId": "deleteRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Request delete"
          },
          "500": {
            "description": "Failed to delete a request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Soft delete a request for the user ",
        "tags": [
          "profiles",
          "requests"
        ]
      }
    },
    "/profiles/requester/requests": {
      "get": {
        "operationId": "getRequests",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "One of the statuses \"open\", \"closed\", or \"pending\"",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicle",
            "required": false,
            "in": "query",
            "description": "The id of the vehicle",
            "schema": {}
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "A text to use to search for a request (fields: description, summary)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Got requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestListReponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get list of requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Get all requests for the user",
        "tags": [
          "profiles",
          "requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shanda API",
    "description": "Shanda API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "data": {
            "type": "object"
          },
          "msg": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "data",
          "msg"
        ]
      },
      "REQUEST_TAGS": {
        "type": "string",
        "enum": [
          "mvi",
          "tires",
          "oil_change",
          "detailing",
          "keys",
          "parts",
          "maintenance"
        ]
      },
      "REQUEST_CATEGORY": {
        "type": "string",
        "enum": [
          "service_request",
          "part_request"
        ]
      },
      "RequestTemplateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "description": "A list of tags to help identify requests",
            "items": {
              "$ref": "#/components/schemas/REQUEST_TAGS"
            }
          },
          "category": {
            "type": "array",
            "description": "A category for requests",
            "items": {
              "$ref": "#/components/schemas/REQUEST_CATEGORY"
            }
          }
        },
        "required": [
          "id",
          "summary",
          "description"
        ]
      },
      "RequestTemplateResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/RequestTemplateDto"
          }
        },
        "description": "Create user data",
        "required": [
          "statusCode",
          "data"
        ]
      },
      "RequestTemplateListReponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequestTemplateDto"
            }
          },
          "count": {
            "type": "number"
          },
          "totalCount": {
            "type": "number"
          }
        },
        "description": "Get request templates",
        "required": [
          "statusCode",
          "data",
          "count",
          "totalCount"
        ]
      },
...
      "CreateRequest": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "vehicleId": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "description": "A list of tags to help identify requests",
            "items": {
              "$ref": "#/components/schemas/REQUEST_TAGS"
            }
          },
          "category": {
            "description": "A category for requests",
            "default": "service_request",
            "allOf": [
              {
                "$ref": "#/components/schemas/REQUEST_CATEGORY"
              }
            ]
          }
        },
        "description": "Create user data",
        "required": [
          "summary",
          "description",
          "vehicleId"
        ]
      },
      "REQUEST_STATUS": {
        "type": "string",
        "enum": [
          "open",
          "closed",
          "pending"
        ]
      },
      "RequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "vehicleId": {
            "type": "string"
          },
          "requesterId": {
            "type": "string"
          },
          "status": {
            "description": "Request status",
            "allOf": [
              {
                "$ref": "#/components/schemas/REQUEST_STATUS"
              }
            ]
          },
          "created": {
            "format": "date-time",
            "type": "string"
          },
          "deleted": {
            "format": "date-time",
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "summary",
          "description",
          "vehicleId",
          "requesterId",
          "status",
          "created",
          "deleted",
          "tags",
          "category"
        ]
      },
      "RequestResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/RequestDto"
          }
        },
        "description": "Create user data",
        "required": [
          "statusCode",
          "data"
        ]
      },
      "UpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "vehicleId": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "description": "A list of tags to help identify requests",
            "items": {
              "$ref": "#/components/schemas/REQUEST_TAGS"
            }
          },
          "category": {
            "type": "array",
            "description": "A category for requests",
            "items": {
              "$ref": "#/components/schemas/REQUEST_CATEGORY"
            }
          }
        },
        "description": "Update request data",
        "required": [
          "id",
          "vehicleId",
          "tags",
          "category"
        ]
      },
      "RequestListReponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequestDto"
            }
          },
          "count": {
            "type": "number"
          },
          "totalCount": {
            "type": "number"
          }
        },
        "description": "Get list of requests",
        "required": [
          "statusCode",
          "data",
          "count",
          "totalCount"
        ]
      }
    }
  }
}